requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.5.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.31.0rc1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.3
requests==2.32.3
rsa==4.9.1
sqlparse==0.5.3
swapper==1.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
urllib3==2.4.0
